remove old auth router, add users router, integrate notes protection and user models

+++ backend/app/api/notes.py
@@ -1,8 +1,8 @@
-from fastapi import APIRouter, HTTPException, Response, Query, Body, Path, Depends
+from fastapi import APIRouter, HTTPException, Response, Query, Body, Path
 from sqlalchemy import select
 
 from app.models import notes_models
-from app.dependencies import SessionDep, oauth2_scheme
+from app.dependencies import SessionDep, UserDep
 from app.db.models.notes_db import Notes
 
 from typing import Annotated
@@ -13,10 +13,10 @@ router = APIRouter(prefix = "/v1/notes")
 @router.get("/", response_model=notes_models.NotesGet)
 async def get_all_notes(
     session: SessionDep, 
-    token: Annotated[str, Depends(oauth2_scheme)],
+    user: UserDep,
     limit: Annotated[int, Query(ge=1, le=100)] = 10,
 ):
-    result = await session.execute(select(Notes).order_by(Notes.created_at).limit(limit))
+    result = await session.execute(select(Notes).order_by(Notes.created_at).limit(limit).where(user.id == Notes.user_id))
     notes = result.scalars().all()
 
     return {"notes": notes}
@@ -24,38 +24,43 @@ async def get_all_notes(
 @router.post("/", response_model=notes_models.NoteResponse)
 async def create_note(
     session: SessionDep, 
-    token: Annotated[str, Depends(oauth2_scheme)],
+    user: UserDep,
     note: Annotated[notes_models.NoteCreate, Body()]
 ):
-    note_orm = Notes(title = note.title, content = note.content)
-    session.add(note_orm)
+    note_orm = Notes(user_id = user.id, title = note.title, content = note.content)
 
+    session.add(note_orm)
     await session.commit()
     await session.refresh(note_orm)
+
     return note_orm
 
 @router.get("/{note_id}", response_model=notes_models.NoteResponse)
 async def get_note(
     session: SessionDep,
-    token: Annotated[str, Depends(oauth2_scheme)], 
+    user: UserDep, 
     note_id: Annotated[uuid.UUID, Path()]
-):
+):  
     result = await session.get(Notes, note_id)
     if not result:
         raise HTTPException(status_code=404, detail="Note not found")
+    if result.user_id != user.id:
+        raise HTTPException(status_code=401, detail="Unauthorized user")
     
     return result
 
 @router.patch("/{note_id}", response_model=notes_models.NoteResponse)
 async def update_note(
     session: SessionDep, 
-    token: Annotated[str, Depends(oauth2_scheme)],
+    user: UserDep,
     note_id: Annotated[uuid.UUID, Path()], 
     note: Annotated[notes_models.NoteUpdate, Body()]
 ):
     result = await session.get(Notes, note_id)
     if not result:
         raise HTTPException(status_code=404, detail="Note not found")
+    if result.user_id != user.id:
+        raise HTTPException(status_code=401, detail="Unauthorized user")
     
     note_details = note.model_dump(exclude_unset=True)
     if not note_details:
@@ -71,12 +76,14 @@ async def update_note(
 @router.delete("/{note_id}")
 async def delete_note(
     session: SessionDep, 
-    token: Annotated[str, Depends(oauth2_scheme)],
+    user: UserDep,
     note_id: Annotated[uuid.UUID, Path()]
 ):
     result = await session.get(Notes, note_id)
     if not result:
         raise HTTPException(status_code=404, detail="Note not found")
+    if result.user_id != user.id:
+        raise HTTPException(status_code=401, detail="Unauthorized user")
     
     await session.delete(result)
     await session.commit()
